starter/src/app/responses/requirements.txt
# FastMCP
fastmcp
fastapi
uvicorn

#DB26ai
oracledb

# Langgraph
langchain_oci
langchain_community
langchain_docling 
langchain_openai
oci-openai
langgraph-cli[inmem]
langchain_mcp_adapters

# OAuth Caching 
aiocache